+++ galsim/chromatic.py
@@ -142,6 +142,30 @@ class ChromaticObject(object):
         """
         ChromaticObject._multiplier_cache.resize(maxsize)
 
+    def _fiducial_profile(self, bandpass):
+        """
+        Return a fiducial achromatic profile of a chromatic object that can be used to estimate
+        default output image characteristics, or in the case of separable profiles, can be scaled to
+        give the monochromatic profile at any wavelength or the wavelength-integrated profile.
+        """
+        bpwave = bandpass.effective_wavelength
+        prof0 = self.evaluateAtWavelength(bpwave)
+        if prof0.flux != 0:
+            return bpwave, prof0
+
+        candidate_waves = np.concatenate(
+            [np.array([0.5 * (bandpass.blue_limit + bandpass.red_limit)]),
+             bandpass.wave_list,
+             self.wave_list])
+        # Prioritize wavelengths near the bandpass effective wavelength.
+        candidate_waves = candidate_waves[np.argsort(np.abs(candidate_waves - bpwave))]
+        for w in candidate_waves:
+            prof0 = self.evaluateAtWavelength(w)
+            if prof0.flux != 0:
+                return w, prof0
+
+        raise ValueError("Could not locate fiducial wavelength where SED * Bandpass is nonzero.")
+
     def __repr__(self):
         return 'galsim.ChromaticObject(%r)'%self.obj
 
@@ -281,8 +305,8 @@ class ChromaticObject(object):
         if self.separable: assert hasattr(self, 'SED')
         assert hasattr(self, 'wave_list')
 
-        # setup output image (semi-arbitrarily using the bandpass effective wavelength)
-        prof0 = self.evaluateAtWavelength(bandpass.effective_wavelength)
+        # setup output image using fiducial profile
+        wave0, prof0 = self._fiducial_profile(bandpass)
         image = prof0.drawImage(image=image, setup_only=True, **kwargs)
         _remove_setup_kwargs(kwargs)
 
@@ -291,7 +315,7 @@ class ChromaticObject(object):
 
         if self.separable:
             multiplier = ChromaticObject._multiplier_cache(self.SED, bandpass, tuple(wave_list))
-            prof0 *= multiplier/self.SED(bandpass.effective_wavelength)
+            prof0 *= multiplier/self.SED(wave0)
             image = prof0.drawImage(image=image, **kwargs)
             return image
 
@@ -880,7 +904,7 @@ class InterpolatedChromaticObject(ChromaticObject):
         # with whatever pixel scale was required to sample all the images properly.  We want to set
         # up an output image that has the requested pixel scale, which might change the image size
         # and so on.
-        prof0 = self.evaluateAtWavelength(bandpass.effective_wavelength)
+        _, prof0 = self._fiducial_profile(bandpass)
         image = prof0.drawImage(image=image, setup_only=True, **kwargs)
         _remove_setup_kwargs(kwargs)
 
@@ -1676,7 +1700,8 @@ class ChromaticConvolution(ChromaticObject):
             SED = lambda w: reduce(lambda x,y:x*y, [s(w) for s in sep_SED], 1)
             insep_obj = galsim.Convolve(insep_profs, gsparams=gsparams)
             # Find scale at which to draw effective profile
-            iiscale = insep_obj.evaluateAtWavelength(bandpass.effective_wavelength).nyquistScale()
+            _, prof0 = insep_obj._fiducial_profile(bandpass)
+            iiscale = prof0.nyquistScale()
             if iimult is not None:
                 iiscale /= iimult
             # Create the effective bandpass.
@@ -1852,7 +1877,7 @@ class ChromaticConvolution(ChromaticObject):
         # and non-ChromaticConvolution).  (The latter case was dealt with in the constructor.)
 
         # setup output image (semi-arbitrarily using the bandpass effective wavelength)
-        prof0 = self.evaluateAtWavelength(bandpass.effective_wavelength)
+        wave0, prof0 = self._fiducial_profile(bandpass)
         image = prof0.drawImage(image=image, setup_only=True, **kwargs)
         _remove_setup_kwargs(kwargs)
 
@@ -1867,8 +1892,8 @@ class ChromaticConvolution(ChromaticObject):
                 if isinstance(obj, galsim.GSObject):
                     sep_profs.append(obj) # The g(x,y)'s (see above)
                 else:
-                    sep_profs.append(obj.evaluateAtWavelength(bandpass.effective_wavelength)
-                                     /obj.SED(bandpass.effective_wavelength)) # more g(x,y)'s
+                    wave0, prof0 = obj._fiducial_profile(bandpass)
+                    sep_profs.append(prof0 / obj.SED(wave0)) # more g(x,y)'s
                     sep_SED.append(obj.SED) # The h(lambda)'s (see above)
                     wave_list = np.union1d(wave_list, obj.wave_list)
             else: